backend/requirements.txt
# backend/requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
Flask-Cors==6.0.1
PyJWT==2.10.1
werkzeug==3.1.3
torch==2.7.1
openai-whisper==20250625
gunicorn==23.0.0
python-docx==1.2.0
fpdf==1.7.2
